converted the Timeline component to be named Lane. This way a Timeline component has children: TimeReference and Lane components
