+++ src/internal.ts
@@ -161,6 +161,7 @@ export {
 
 // Tissue (2)
 import { Navigation } from './components/Navigation/Navigation'
+import { Lane } from './components/Timeline/Lane'
 import { Location } from './components/Location/Location'
 import { TimePicker } from './components/TimePicker/TimePicker'
 import { DatePicker } from './components/DatePicker/DatePicker'
@@ -170,6 +171,7 @@ import { TimeZone } from './components/TimeZone/TimeZone'
 import { LabelEditor } from './components/LabelEditor/LabelEditor'
 export {
   Navigation,
+  Lane,
   Location,
   TimePicker,
   DatePicker,
@@ -180,11 +182,11 @@ export {
 }
 
 // Organs (3)
-import { Timeline } from './components/Timeline/Timeline'
 import { LabelManager } from './components/LabelManager/LabelManager'
+import { Timeline } from './components/Timeline/Timeline'
 export {
-  Timeline,
-  LabelManager
+  LabelManager,
+  Timeline
 }
 
 // Organisms (4)
